Include additional files in `attachments`
Fixes attachments not included in edit messages, causing not found error responses. Note that this looks like a Discord bug.

For CV2 attachments, this isn't useful when sending messages as the descriptions are already in the component itself

But this will be useful later when embed images benefit from the same "auto uploads" as CV2